pyproject.toml
[tool.poetry]
name = "jjbdiscordbot"
version = "0.1.0"
description = "Discord bot for the JJB Discord Bot"
authors = ["Josh Jacobs <josh@dangerous.tech>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
